Correcting spelling errors in main.py

+++ main.py
@@ -301,10 +301,11 @@ def delete_folder_contents(folder_path):
     for filename in os.listdir(folder_path):
         file_path = os.path.join(folder_path, filename)
         try:
-            if os.path.isfile(file_path) or os.path.islink(file_path):
-                os.unlink(file_path)
-            elif os.path.isdir(file_path):
-                shutil.rmtree(file_path)
+            if filename != ".gitkeep":
+                if os.path.isfile(file_path) or os.path.islink(file_path):
+                    os.unlink(file_path)
+                elif os.path.isdir(file_path):
+                    shutil.rmtree(file_path)
         except Exception as e:
             print(f"Failed to delete {file_path}. Reason: {e}")
            
@@ -335,7 +336,7 @@ def transcribe():
     print(
     """
     NOTE:
-    If you transcribed resaerch data, MAKE SURE TO DELETE THE WHISPER MODEL before turning the internet back on.
+    If you transcribed research data, MAKE SURE TO DELETE THE WHISPER MODEL before turning the internet back on.
     You can find the model(s) you downloaded in the 'model' folder in the 'EasyWhisper' folder.
     Please delete these models before connecting your computer to the internet again.
     
@@ -376,7 +377,7 @@ def diarize():
     print(
     """
     NOTE:
-    If you transcribed resaerch data, MAKE SURE TO DELETE THE WHISPER MODEL before turning the internet back on.
+    If you transcribed research data, MAKE SURE TO DELETE THE WHISPER MODEL before turning the internet back on.
     You can find the model(s) you downloaded in the 'model' folder in the 'EasyWhisper' folder.
     Please delete these models before connecting your computer to the internet again.
